requirements.txt
google-generativeai
streamlit
trafilatura
